move script into drivers since we wanna package it
